Fix calls to howm-template-string in howm-create-file-with-title
When creating a new note, `howm-insert-template' is called twice to
create the second candidate (the version without %file) accessible
using undo.

This behavior causes variable `howm-template' to also be
called/referred twice.  This seems harmless but if the user customizes
this variable and is defined as a function, there can be unintended
side effects.  For instance, if the custom function does a
completing-read to ask for a template to use, the prompt will show
twice.

To fix, move call to `howm-template-string' outside of the lambda
function in `howm-create-file-with-title' and then use the resulting
value inside of the lambda.